Member DTO의 LoginFailuerCount getter setter 메서드를 short 대신 int로 타입 형변환 하여 get set 하도록 변경

+++ member-service/src/main/java/com/example/member/Member.java
@@ -50,11 +50,11 @@ public class Member {
 	public void setPhone(String phone) {
 		this.phone = phone;
 	}
-	public short getLoginFailureCount() {
+	public int getLoginFailureCount() {
 		return loginFailureCount;
 	}
-	public void setLoginFailureCount(short loginFailureCount) {
-		this.loginFailureCount = loginFailureCount;
+	public void setLoginFailureCount(int loginFailureCount) {
+		this.loginFailureCount = (short)loginFailureCount;
 	}
 	public Date getLastLoginDate() {
 		return lastLoginDate;
